requirements.txt
envparse==0.2.0
multicolorcaptcha
faker
unidecode
pyromod==1.5
ffmpeg-python
humanize
wget==3.2
secureme==0.0.4
cowpy
motor==3.3.1
gpytranslate==1.3.0
GitPython==3.1.29
emoji==1.6.3
pretty-errors==1.2.25
lxml
CloudScraper
textblob==0.17.1
beautifulsoup4
requests
lexica-api
sqlalchemy==1.3.23
telethon==1.24.0
python-telegram-bot==13.13
pyrogram
psycopg2-binary
feedparser
pynewtonmath
spongemock
zalgo-text
geopy
redis
psutil
jikanpy
speedtest-cli==2.1.3
coffeehouse
regex==2022.10.31
bleach
wikipedia
spamwatch
alphabet_detector
cachetools
dnspython==2.1.0
fontTools
gtts
tswift
pytz
PyYAML==6.0
html2text
pythonping
html_telegraph_poster
pytube
youtube-search-python
PyLyrics
hachoir
pymongo
pybase64
pySmartDL
validators
asyncio
nltk
aiohttp==3.9.5
pykeyboard==0.1.2
pillow
CurrencyConverter
googletrans==4.0.0-rc1
google_trans_new
pyrate_limiter==2.10.0
nekos.py
tgcrypto
python-arq==6.0.5
hachoir
configparser
python-dotenv==0.19.2
bing_image_downloader
parsel
urllib3
heroku3
telegraph
youtube_search
search_engine_parser
glitch_this
NumPy
opencv-python-headless
countryinfo==0.1.2
python-rapidjson==1.9
pytimeparse
markdown2
pycryptodome
base58
httpx
httpx[http2]
async-timeout==4.0.2
loguru
blackpink
